+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
  - Node.js v14 support.
  - `steadybit service init` gives more meaningful error messages in case the Steadybit servers fails to respond successfully with teams.
+ - The service definition file name may now be supplied via the `-f` option. By default, all commands assume that the default service definition file name is `.steadybit.yml`.
 
 ## v2.1.1