+++ CHANGELOG.md
@@ -4,7 +4,8 @@
  - Node.js v14 support.
  - `steadybit service init` gives more meaningful error messages in case the Steadybit servers fails to respond successfully with teams.
  - The service definition file name may now be supplied via the `-f` option. By default, all commands assume that the default service definition file name is `.steadybit.yml`.
- - Print a success message when a service definition was applied
+ - Print a success message when a service definition was applied.
+ - Change label from `pending` to `not checked` in `steadybit service verify` output.
 
 ## v2.1.1